\name{CPHshape-internal}
\alias{CPHshape-internal}
\alias{find.antimodeFULLMLE}
\alias{find.beta}
\alias{find.decreasingFULLMLE}
\alias{find.increasingFULLMLE}
\alias{find.modeFULLMLE}
\alias{find.unimodalFULLMLE}
\alias{find.h.up}
\alias{find.h.down}
\alias{find.increasingMLE}
\alias{find.decreasingMLE}
\alias{find.modeMLE}
\alias{find.unimodalMLE}
\alias{find.antimodeMLE}
\alias{find.ushapedMLE}
\title{Internal functions for CPHshape.}
\description{Internal functions for CPHshape.}
\details{These are not to be called by the user.}
\keyword{internal}